Add a setting to control enqueueing new requests
If server.enqueue_enabled=false, the requests are not put into
the postgresql database. Insread they are proxied to the destination
and the response (whether 200 OK or Bad Request 400) is returned.

This setting can be used to control database load if latency is too
high. But it still requires a database to connect.

Signed-off-by: Valentin Kiselev <mrexox@evilmartians.com>